# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OrgFlow API
  version: 1.0.0
  description: Employee and leave management. All routes except login and password reset need a Bearer JWT.
servers:
  - url: http://localhost:8080/api
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
...
security:
  - bearer: []
paths:
...
  /employees:
    get: { summary: List with search, departmentId, status, page, limit (salary visible to admins only), responses: { '200': { description: paged list } } }
    post: { summary: Create employee, optionally with a login (admin), responses: { '201': { description: created }, '409': { description: duplicate email } } }
...
  /employees/import:
    post:
      summary: "Bulk-create employees from CSV text (admin). Always 200: check errors[]. Nothing is saved unless errors is empty; dryRun true only validates."
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [csv], properties: { csv: { type: string }, dryRun: { type: boolean } } } } } }
      responses: { '200': { description: "{ total, valid, created, errors: [{ row, field, message }] }" }, '400': { description: unreadable file or missing columns } }
...
  /employees/{id}:
    parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
    get: { summary: One employee, responses: { '200': { description: employee } } }
    put: { summary: Update employee (admin), responses: { '200': { description: updated } } }
    delete: { summary: Deactivate employee and disable sign-in (admin), responses: { '204': { description: deactivated } } }
...
  /holidays:
    get: { summary: Company holidays, optional year, responses: { '200': { description: list } } }
    post: { summary: Add a holiday (admin), responses: { '201': { description: created }, '409': { description: date already exists } } }
  /holidays/{id}:
    parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
    delete: { summary: Remove a holiday (admin), responses: { '204': { description: removed } } }
...